Added IMU Position Tracking and PID for PPS and Vision Sensor Auto Pick Up
IMU Position Tracking has been implemented. Not sure if it will work, still needs to be tested. Also added custom PIDs for both pure pursuit and visionSensor picking up discs so hopefully those are smoother. Commented out TBH for the flywheel, it needs more testing and I am not sure how effective it actually is.  Also added a second controller code for automated functions in driver control like driveToRoller or TurnToHighGoal. Also started on a new autonomous that is simpler and should be more consistent.